requirements.txt
discord~=2.3.2
qq-botpy~=1.2.1

mysql-connector-python~=8.3.0
python-a2s~=1.3.0
requests~=2.31.0
pycountry~=23.12.11
rcon~=2.4.6

python-dotenv~=1.0.0
requests~=2.31.0
pycountry~=23.12.11
pytz~=2024.1
python-steam-api~=1.2.2
beautifulsoup4~=4.12.3
tqdm>=4.66.2

steam~=1.4.4
selenium~=4.19.0
pillow~=10.2.0
rsp~=1.0.1
